Add boundPath clientTls module. boundPath is a client tls module that allows you to determine the common name based on the destination bound path.
