+++ linkerd/tls/src/main/resources/META-INF/services/io.buoyant.linkerd.TlsClientInitializer
@@ -1,2 +1,3 @@
 io.l5d.clientTls.noValidation
 io.l5d.clientTls.static
+io.l5d.clientTls.boundPath
